requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymongo==4.3.3
rasa==3.6.21
faiss-cpu==1.8.0.post1
sentence-transformers==3.4.1
python-dotenv==1.0.1
requests==2.32.3